#221 Share rmp work
