+++ src/components/menu/account-view/saves-section.tsx
@@ -3,7 +3,7 @@ import { RMSection, RMSectionBody, RMSectionHeader } from '@railmapgen/mantine-c
 import { logger } from '@railmapgen/rmg-runtime';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { MdDeleteOutline, MdOutlineSync, MdOutlineSyncAlt } from 'react-icons/md';
+import { MdDeleteOutline, MdOutlineShare, MdOutlineSync, MdOutlineSyncAlt } from 'react-icons/md';
 import { useRootDispatch, useRootSelector } from '../../../redux';
 import { fetchSaveList, logout, syncAfterLogin } from '../../../redux/account/account-slice';
 import { addNotification } from '../../../redux/notification/notification-slice';
@@ -11,7 +11,9 @@ import { setLastChangedAtTimeStamp } from '../../../redux/rmp-save/rmp-save-slic
 import { apiFetch } from '../../../util/api';
 import { API_ENDPOINT, APISaveList, SAVE_KEY } from '../../../util/constants';
 import { getRMPSave, notifyRMPSaveChange, setRMPSave } from '../../../util/local-storage-save';
+import { getRandomCityName } from '../../../util/random-save-name';
 import InlineEdit from '../../common/inline-edit';
+import ShareModal from './share-modal';
 
 const MAXIMUM_FREE_SAVE = 1;
 const MAXIMUM_SAVE = 10;
@@ -38,6 +40,8 @@ const SavesSection = () => {
 
     const [syncButtonIsLoading, setSyncButtonIsLoading] = React.useState(undefined as number | undefined);
     const [deleteButtonIsLoading, setDeleteButtonIsLoading] = React.useState(undefined as number | undefined);
+    const [shareModalOpened, setShareModalOpened] = React.useState(false);
+    const [shareSaveID, setShareSaveID] = React.useState<number | undefined>(undefined);
 
     const showErrorToast = (msg: string) =>
         dispatch(
@@ -56,7 +60,7 @@ const SavesSection = () => {
             return;
         }
         const { data, hash } = save;
-        const index = getRandomCity();
+        const index = getRandomCityName();
         try {
             const rep = await apiFetch(
                 API_ENDPOINT.SAVES,
@@ -206,6 +210,11 @@ const SavesSection = () => {
         dispatch(fetchSaveList());
     };
 
+    const handleOpenShareModal = (saveID: number) => {
+        setShareSaveID(saveID);
+        setShareModalOpened(true);
+    };
+
     return (
         <RMSection>
             <RMSectionHeader align="center">
@@ -224,7 +233,7 @@ const SavesSection = () => {
                             <InlineEdit
                                 initialValue={_.index}
                                 onSave={val => handleEditSaveName(_.id, val)}
-                                textInputWidth="207px"
+                                textInputWidth="177px"
                             />
                             <ActionIcon
                                 disabled={isUpdateDisabled(_.id)}
@@ -233,6 +242,9 @@ const SavesSection = () => {
                                 title={_.id === currentSaveId ? t('Sync now') : t('Sync this slot')}
                             >
                                 {_.id === currentSaveId ? <MdOutlineSync /> : <MdOutlineSyncAlt />}
+                            </ActionIcon>
+                            <ActionIcon onClick={() => handleOpenShareModal(_.id)} title={t('Share')}>
+                                <MdOutlineShare />
                             </ActionIcon>
                             <ActionIcon
                                 color="red"
@@ -257,397 +269,13 @@ const SavesSection = () => {
                     </Card>
                 ))}
             </RMSectionBody>
+            <ShareModal
+                opened={shareModalOpened}
+                onClose={() => setShareModalOpened(false)}
+                shareSaveInfo={saveList?.find(s => s.id === shareSaveID)}
+            />
         </RMSection>
     );
 };
 
 export default SavesSection;
-
-const CITIES_LIST = [
-    'Tokyo',
-    'Paris',
-    'London',
-    'New York',
-    'Shanghai',
-    'Dubai',
-    'Sydney',
-    'Rio de Janeiro',
-    'Cairo',
-    'Moscow',
-    'Rome',
-    'Berlin',
-    'Madrid',
-    'Seoul',
-    'Singapore',
-    'Hong Kong',
-    'Beijing',
-    'Mexico City',
-    'Buenos Aires',
-    'Istanbul',
-    'Lagos',
-    'Jakarta',
-    'Osaka',
-    'Manila',
-    'Chongqing',
-    'Karachi',
-    'Dhaka',
-    'Lima',
-    'Bangkok',
-    'Lahore',
-    'Ho Chi Minh City',
-    'Shenzhen',
-    'Tianjin',
-    'Taipei',
-    'Bogota',
-    'Baghdad',
-    'Riyadh',
-    'Santiago',
-    'Guangzhou',
-    'Yangon',
-    'Johannesburg',
-    'Toronto',
-    'Montreal',
-    'Vancouver',
-    'Sao Paulo',
-    'Kuala Lumpur',
-    'Athens',
-    'Lisbon',
-    'Dublin',
-    'Amsterdam',
-    'Vienna',
-    'Prague',
-    'Budapest',
-    'Warsaw',
-    'Stockholm',
-    'Oslo',
-    'Helsinki',
-    'Copenhagen',
-    'Brussels',
-    'Geneva',
-    'Zurich',
-    'Barcelona',
-    'Milan',
-    'Florence',
-    'Venice',
-    'Munich',
-    'Hamburg',
-    'Frankfurt',
-    'Kyoto',
-    'Nairobi',
-    'Cape Town',
-    'Casablanca',
-    'Marrakesh',
-    'Accra',
-    'Addis Ababa',
-    'Alexandria',
-    'Ankara',
-    'Jerusalem',
-    'Tel Aviv',
-    'Beirut',
-    'Abu Dhabi',
-    'Doha',
-    'Kuwait City',
-    'Muscat',
-    'Amman',
-    'Baku',
-    'Tbilisi',
-    'Yerevan',
-    'Almaty',
-    'Tashkent',
-    'Kabul',
-    'Tehran',
-    'Damascus',
-    "Sana'a",
-    'Aden',
-    'Manama',
-    'Jeddah',
-    'Mecca',
-    'Medina',
-    'Salalah',
-    'Nicosia',
-    'Thessaloniki',
-    'Sofia',
-    'Bucharest',
-    'Belgrade',
-    'Zagreb',
-    'Sarajevo',
-    'Skopje',
-    'Tirana',
-    'Pristina',
-    'Podgorica',
-    'Ljubljana',
-    'Bratislava',
-    'Krakow',
-    'Gdansk',
-    'Wroclaw',
-    'Poznan',
-    'Vilnius',
-    'Riga',
-    'Tallinn',
-    'Reykjavik',
-    'Manchester',
-    'Birmingham',
-    'Glasgow',
-    'Edinburgh',
-    'Cardiff',
-    'Belfast',
-    'Marseille',
-    'Lyon',
-    'Toulouse',
-    'Nice',
-    'Nantes',
-    'Strasbourg',
-    'Bordeaux',
-    'Lille',
-    'Cologne',
-    'Stuttgart',
-    'Dusseldorf',
-    'Dortmund',
-    'Essen',
-    'Leipzig',
-    'Dresden',
-    'Hanover',
-    'Nuremberg',
-    'Bremen',
-    'Rotterdam',
-    'The Hague',
-    'Utrecht',
-    'Eindhoven',
-    'Porto',
-    'Valencia',
-    'Seville',
-    'Zaragoza',
-    'Malaga',
-    'Murcia',
-    'Palma de Mallorca',
-    'Las Palmas de Gran Canaria',
-    'Antwerp',
-    'Ghent',
-    'Charleroi',
-    'Liège',
-    'Bern',
-    'Basel',
-    'Lausanne',
-    'Graz',
-    'Linz',
-    'Salzburg',
-    'Innsbruck',
-    'Naples',
-    'Turin',
-    'Palermo',
-    'Genoa',
-    'Bologna',
-    'Bari',
-    'Catania',
-    'Verona',
-    'Melbourne',
-    'Brisbane',
-    'Perth',
-    'Adelaide',
-    'Gold Coast',
-    'Canberra',
-    'Auckland',
-    'Wellington',
-    'Christchurch',
-    'Dunedin',
-    'Suva',
-    'Port Moresby',
-    'Noumea',
-    'Papeete',
-    'Honolulu',
-    'Quezon City',
-    'Davao City',
-    'Cebu City',
-    'Chiang Mai',
-    'Phuket',
-    'George Town',
-    'Ipoh',
-    'Johor Bahru',
-    'Surabaya',
-    'Bandung',
-    'Medan',
-    'Semarang',
-    'Santiago',
-    'Lima',
-    'Bogota',
-    'Caracas',
-    'Quito',
-    'La Paz',
-    'Sucre',
-    'Buenos Aires',
-    'Cordoba',
-    'Rosario',
-    'Mendoza',
-    'Valparaiso',
-    'Concepcion',
-    'Arequipa',
-    'Trujillo',
-    'Chiclayo',
-    'Piura',
-    'Medellin',
-    'Cali',
-    'Barranquilla',
-    'Cartagena',
-    'Maracaibo',
-    'Valencia',
-    'Barquisimeto',
-    'Ciudad Guayana',
-    'Guayaquil',
-    'Cuenca',
-    'Santo Domingo',
-    'Santa Cruz de la Sierra',
-    'Cochabamba',
-    'Guadalajara',
-    'Monterrey',
-    'Puebla',
-    'Tijuana',
-    'Leon',
-    'Juarez',
-    'Ecatepec',
-    'Zapopan',
-    'Nezahualcoyotl',
-    'Chihuahua',
-    'Naucalpan',
-    'Merida',
-    'Aguascalientes',
-    'San Luis Potosi',
-    'Hermosillo',
-    'Saltillo',
-    'Mexicali',
-    'Culiacan',
-    'Acapulco',
-    'Guadalupe',
-    'Cancun',
-    'Torreon',
-    'El Paso',
-    'Tucson',
-    'Phoenix',
-    'Los Angeles',
-    'San Diego',
-    'San Jose',
-    'San Francisco',
-    'Seattle',
-    'Portland',
-    'Las Vegas',
-    'Denver',
-    'Chicago',
-    'Houston',
-    'Dallas',
-    'San Antonio',
-    'Austin',
-    'Philadelphia',
-    'Washington D.C.',
-    'Boston',
-    'Miami',
-    'Atlanta',
-    'Orlando',
-    'Tampa',
-    'Detroit',
-    'Cleveland',
-    'Minneapolis',
-    'St. Louis',
-    'Kansas City',
-    'New Orleans',
-    'Charlotte',
-    'Nashville',
-    'Indianapolis',
-    'Columbus',
-    'Cincinnati',
-    'Pittsburgh',
-    'Milwaukee',
-    'Calgary',
-    'Edmonton',
-    'Winnipeg',
-    'Ottawa',
-    'Quebec City',
-    'Halifax',
-    'Chengdu',
-    'Wuhan',
-    'Hangzhou',
-    'Suzhou',
-    'Xian',
-    'Nanjing',
-    'Zhengzhou',
-    'Jinan',
-    'Qingdao',
-    'Hefei',
-    'Fuzhou',
-    'Xiamen',
-    'Ningbo',
-    'Wuxi',
-    'Kunming',
-    'Changsha',
-    'Nanchang',
-    'Shijiazhuang',
-    'Harbin',
-    'Changchun',
-    'Taiyuan',
-    'Urumqi',
-    'Lanzhou',
-    'Yinchuan',
-    'Xining',
-    'Guiyang',
-    'Lhasa',
-    'Hohhot',
-    'Nanning',
-    'Haikou',
-    'Sanya',
-    'Zhuhai',
-    'Foshan',
-    'Dongguan',
-    'Zhongshan',
-    'Shantou',
-    'Wenzhou',
-    'Jiaxing',
-    'Shaoxing',
-    'Taizhou',
-    'Jinhua',
-    'Quanzhou',
-    'Putian',
-    'Zhangzhou',
-    'Longyan',
-    'Sanming',
-    'Nanping',
-    'Yingtan',
-    'Shangrao',
-    'Jiujiang',
-    'Jingdezhen',
-    'Pingxiang',
-    'Xinyu',
-    'Yichun',
-    'Ganzhou',
-    'Qingyuan',
-    'Shaoguan',
-    'Heyuan',
-    'Meizhou',
-    'Shanwei',
-    'Yangjiang',
-    'Zhanjiang',
-    'Maoming',
-    'Zhaoqing',
-    'Yunfu',
-    'Laibin',
-    'Chongzuo',
-    'Hechi',
-    'Baise',
-    'Wuzhou',
-    'Fangchenggang',
-    'Qinzhou',
-    'Beihai',
-    'Guigang',
-    'Yulin',
-    'Cenxi',
-    'Dongxing',
-    'Anshan',
-    'Fushun',
-    'Jilin City',
-    'Qiqihar',
-    'Daqing',
-    'Mudanjiang',
-    'Jixi',
-];
-
-const getRandomCity = () => {
-    const randomIndex = Math.floor(Math.random() * CITIES_LIST.length);
-    return CITIES_LIST[randomIndex];
-};
