+++ src/i18n/translations/ja.json
@@ -23,6 +23,12 @@
         "text4": "ご不明な場合、または当社の開発者の指示なしに操作している場合は、このダイアログを閉じてください。"
     },
 
+    "Session expired.": "セッションの有効期限が切れました。",
+    "Please log in again.": "再度ログインしてください。",
+    "Error": "錯誤",
+    "Success": "成功",
+    "Close": "關閉",
+
     "About": "概要",
     "Allow cookies to help improve our website": "Cookie を許可して網頁の改善にご協力ください",
     "Appearance": "外観",
@@ -164,5 +170,17 @@
     "Replace cloud with local": "雲を本地で置き換える",
     "Download Local save": "本地保存をダウンロード",
     "Replace local with cloud": "本地を雲で置き換える",
-    "Download Cloud save": "雲保存をダウンロード"
+    "Download Cloud save": "雲保存をダウンロード",
+
+    "Share work {{index}}": "作品 {{index}} を共有",
+    "Share this link with others to allow them to view this work.": "このリンクを他の方と共有し、作品の閲覧を許可してください。",
+    "Copied": "コピーしました",
+    "Copy": "コピー",
+    "Share link created successfully.": "共有リンクが正常に作成されました。",
+    "Share link deleted successfully.": "共有リンクが正常に削除されました。",
+    "This link is valid until:": "このリンクの有効期限：",
+    "Delete Share": "共有を削除",
+    "Create Share Link": "共有リンクを作成",
+    "Please be aware that sharing this link does not create a separate copy of your work. Instead, everyone who opens the link will always see the most up-to-date version. So, if you make changes after sharing the link, those changes will be visible to everyone who has the link.": "ご注意：このリンクを共有しても、作品の個別のコピーは作成されません。代わりに、リンクを開いた人は常に最新バージョンを閲覧することになります。そのため、リンク共有後に変更を加えた場合、その変更はリンクをお持ちのすべての方に表示されます。",
+    "Link will be valid as long as you have an active subscription. Otherwise, the link will be valid for 1 day after creating it.": "有効なご購読がある限り、リンクは有効です。それ以外の場合、リンクは作成後1日間有効となります。"
 }
