+++ src/i18n/translations/zh-Hans.json
@@ -23,6 +23,12 @@
         "text4": "如果您不确定，或者不是在我们的开发者的指引下操作，请关闭这个对话框。"
     },
 
+    "Session expired.": "会话已过期。",
+    "Please log in again.": "请重新登录。",
+    "Error": "错误",
+    "Success": "成功",
+    "Close": "关闭",
+
     "About": "关于",
     "Allow cookies to help improve our website": "允许用 Cookies 帮助改进本网站",
     "Appearance": "外观",
@@ -164,5 +170,17 @@
     "Replace cloud with local": "用本地存档替换云端",
     "Download Local save": "下载本地存档",
     "Replace local with cloud": "用云端存档替换本地",
-    "Download Cloud save": "下载云端存档"
+    "Download Cloud save": "下载云端存档",
+
+    "Share work {{index}}": "分享作品 {{index}}",
+    "Share this link with others to allow them to view this work.": "分享此链接给他人以查看此作品。",
+    "Copied": "已复制",
+    "Copy": "复制",
+    "Share link created successfully.": "分享链接创建成功。",
+    "Share link deleted successfully.": "分享链接删除成功。",
+    "This link is valid until:": "此链接有效期至：",
+    "Delete Share": "删除分享",
+    "Create Share Link": "创建分享链接",
+    "Please be aware that sharing this link does not create a separate copy of your work. Instead, everyone who opens the link will always see the most up-to-date version. So, if you make changes after sharing the link, those changes will be visible to everyone who has the link.": "请注意，分享此链接不会创建您作品的副本。相反，每个打开链接的人都将始终看到最新版本。因此，如果您在分享链接后进行更改，这些更改将对拥有该链接的每个人可见。",
+    "Link will be valid as long as you have an active subscription. Otherwise, the link will be valid for 1 day after creating it.": "只要您拥有有效订阅，链接便会一直有效。否则，链接将在创建后 1 天内有效。"
 }
