+++ src/util/random-save-name.ts
@@ -1,0 +1,388 @@
+const CITIES_LIST = [
+    'Tokyo',
+    'Paris',
+    'London',
+    'New York',
+    'Shanghai',
+    'Dubai',
+    'Sydney',
+    'Rio de Janeiro',
+    'Cairo',
+    'Moscow',
+    'Rome',
+    'Berlin',
+    'Madrid',
+    'Seoul',
+    'Singapore',
+    'Hong Kong',
+    'Beijing',
+    'Mexico City',
+    'Buenos Aires',
+    'Istanbul',
+    'Lagos',
+    'Jakarta',
+    'Osaka',
+    'Manila',
+    'Chongqing',
+    'Karachi',
+    'Dhaka',
+    'Lima',
+    'Bangkok',
+    'Lahore',
+    'Ho Chi Minh City',
+    'Shenzhen',
+    'Tianjin',
+    'Taipei',
+    'Bogota',
+    'Baghdad',
+    'Riyadh',
+    'Santiago',
+    'Guangzhou',
+    'Yangon',
+    'Johannesburg',
+    'Toronto',
+    'Montreal',
+    'Vancouver',
+    'Sao Paulo',
+    'Kuala Lumpur',
+    'Athens',
+    'Lisbon',
+    'Dublin',
+    'Amsterdam',
+    'Vienna',
+    'Prague',
+    'Budapest',
+    'Warsaw',
+    'Stockholm',
+    'Oslo',
+    'Helsinki',
+    'Copenhagen',
+    'Brussels',
+    'Geneva',
+    'Zurich',
+    'Barcelona',
+    'Milan',
+    'Florence',
+    'Venice',
+    'Munich',
+    'Hamburg',
+    'Frankfurt',
+    'Kyoto',
+    'Nairobi',
+    'Cape Town',
+    'Casablanca',
+    'Marrakesh',
+    'Accra',
+    'Addis Ababa',
+    'Alexandria',
+    'Ankara',
+    'Jerusalem',
+    'Tel Aviv',
+    'Beirut',
+    'Abu Dhabi',
+    'Doha',
+    'Kuwait City',
+    'Muscat',
+    'Amman',
+    'Baku',
+    'Tbilisi',
+    'Yerevan',
+    'Almaty',
+    'Tashkent',
+    'Kabul',
+    'Tehran',
+    'Damascus',
+    "Sana'a",
+    'Aden',
+    'Manama',
+    'Jeddah',
+    'Mecca',
+    'Medina',
+    'Salalah',
+    'Nicosia',
+    'Thessaloniki',
+    'Sofia',
+    'Bucharest',
+    'Belgrade',
+    'Zagreb',
+    'Sarajevo',
+    'Skopje',
+    'Tirana',
+    'Pristina',
+    'Podgorica',
+    'Ljubljana',
+    'Bratislava',
+    'Krakow',
+    'Gdansk',
+    'Wroclaw',
+    'Poznan',
+    'Vilnius',
+    'Riga',
+    'Tallinn',
+    'Reykjavik',
+    'Manchester',
+    'Birmingham',
+    'Glasgow',
+    'Edinburgh',
+    'Cardiff',
+    'Belfast',
+    'Marseille',
+    'Lyon',
+    'Toulouse',
+    'Nice',
+    'Nantes',
+    'Strasbourg',
+    'Bordeaux',
+    'Lille',
+    'Cologne',
+    'Stuttgart',
+    'Dusseldorf',
+    'Dortmund',
+    'Essen',
+    'Leipzig',
+    'Dresden',
+    'Hanover',
+    'Nuremberg',
+    'Bremen',
+    'Rotterdam',
+    'The Hague',
+    'Utrecht',
+    'Eindhoven',
+    'Porto',
+    'Valencia',
+    'Seville',
+    'Zaragoza',
+    'Malaga',
+    'Murcia',
+    'Palma de Mallorca',
+    'Las Palmas de Gran Canaria',
+    'Antwerp',
+    'Ghent',
+    'Charleroi',
+    'Liège',
+    'Bern',
+    'Basel',
+    'Lausanne',
+    'Graz',
+    'Linz',
+    'Salzburg',
+    'Innsbruck',
+    'Naples',
+    'Turin',
+    'Palermo',
+    'Genoa',
+    'Bologna',
+    'Bari',
+    'Catania',
+    'Verona',
+    'Melbourne',
+    'Brisbane',
+    'Perth',
+    'Adelaide',
+    'Gold Coast',
+    'Canberra',
+    'Auckland',
+    'Wellington',
+    'Christchurch',
+    'Dunedin',
+    'Suva',
+    'Port Moresby',
+    'Noumea',
+    'Papeete',
+    'Honolulu',
+    'Quezon City',
+    'Davao City',
+    'Cebu City',
+    'Chiang Mai',
+    'Phuket',
+    'George Town',
+    'Ipoh',
+    'Johor Bahru',
+    'Surabaya',
+    'Bandung',
+    'Medan',
+    'Semarang',
+    'Santiago',
+    'Lima',
+    'Bogota',
+    'Caracas',
+    'Quito',
+    'La Paz',
+    'Sucre',
+    'Buenos Aires',
+    'Cordoba',
+    'Rosario',
+    'Mendoza',
+    'Valparaiso',
+    'Concepcion',
+    'Arequipa',
+    'Trujillo',
+    'Chiclayo',
+    'Piura',
+    'Medellin',
+    'Cali',
+    'Barranquilla',
+    'Cartagena',
+    'Maracaibo',
+    'Valencia',
+    'Barquisimeto',
+    'Ciudad Guayana',
+    'Guayaquil',
+    'Cuenca',
+    'Santo Domingo',
+    'Santa Cruz de la Sierra',
+    'Cochabamba',
+    'Guadalajara',
+    'Monterrey',
+    'Puebla',
+    'Tijuana',
+    'Leon',
+    'Juarez',
+    'Ecatepec',
+    'Zapopan',
+    'Nezahualcoyotl',
+    'Chihuahua',
+    'Naucalpan',
+    'Merida',
+    'Aguascalientes',
+    'San Luis Potosi',
+    'Hermosillo',
+    'Saltillo',
+    'Mexicali',
+    'Culiacan',
+    'Acapulco',
+    'Guadalupe',
+    'Cancun',
+    'Torreon',
+    'El Paso',
+    'Tucson',
+    'Phoenix',
+    'Los Angeles',
+    'San Diego',
+    'San Jose',
+    'San Francisco',
+    'Seattle',
+    'Portland',
+    'Las Vegas',
+    'Denver',
+    'Chicago',
+    'Houston',
+    'Dallas',
+    'San Antonio',
+    'Austin',
+    'Philadelphia',
+    'Washington D.C.',
+    'Boston',
+    'Miami',
+    'Atlanta',
+    'Orlando',
+    'Tampa',
+    'Detroit',
+    'Cleveland',
+    'Minneapolis',
+    'St. Louis',
+    'Kansas City',
+    'New Orleans',
+    'Charlotte',
+    'Nashville',
+    'Indianapolis',
+    'Columbus',
+    'Cincinnati',
+    'Pittsburgh',
+    'Milwaukee',
+    'Calgary',
+    'Edmonton',
+    'Winnipeg',
+    'Ottawa',
+    'Quebec City',
+    'Halifax',
+    'Chengdu',
+    'Wuhan',
+    'Hangzhou',
+    'Suzhou',
+    'Xian',
+    'Nanjing',
+    'Zhengzhou',
+    'Jinan',
+    'Qingdao',
+    'Hefei',
+    'Fuzhou',
+    'Xiamen',
+    'Ningbo',
+    'Wuxi',
+    'Kunming',
+    'Changsha',
+    'Nanchang',
+    'Shijiazhuang',
+    'Harbin',
+    'Changchun',
+    'Taiyuan',
+    'Urumqi',
+    'Lanzhou',
+    'Yinchuan',
+    'Xining',
+    'Guiyang',
+    'Lhasa',
+    'Hohhot',
+    'Nanning',
+    'Haikou',
+    'Sanya',
+    'Zhuhai',
+    'Foshan',
+    'Dongguan',
+    'Zhongshan',
+    'Shantou',
+    'Wenzhou',
+    'Jiaxing',
+    'Shaoxing',
+    'Taizhou',
+    'Jinhua',
+    'Quanzhou',
+    'Putian',
+    'Zhangzhou',
+    'Longyan',
+    'Sanming',
+    'Nanping',
+    'Yingtan',
+    'Shangrao',
+    'Jiujiang',
+    'Jingdezhen',
+    'Pingxiang',
+    'Xinyu',
+    'Yichun',
+    'Ganzhou',
+    'Qingyuan',
+    'Shaoguan',
+    'Heyuan',
+    'Meizhou',
+    'Shanwei',
+    'Yangjiang',
+    'Zhanjiang',
+    'Maoming',
+    'Zhaoqing',
+    'Yunfu',
+    'Laibin',
+    'Chongzuo',
+    'Hechi',
+    'Baise',
+    'Wuzhou',
+    'Fangchenggang',
+    'Qinzhou',
+    'Beihai',
+    'Guigang',
+    'Yulin',
+    'Cenxi',
+    'Dongxing',
+    'Anshan',
+    'Fushun',
+    'Jilin City',
+    'Qiqihar',
+    'Daqing',
+    'Mudanjiang',
+    'Jixi',
+];
+
+export const getRandomCityName = () => {
+    const randomIndex = Math.floor(Math.random() * CITIES_LIST.length);
+    return CITIES_LIST[randomIndex];
+};
